ucetni-whatsapp-bot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.9.1
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosmtplib==3.0.1
aiosqlite==0.19.0
alembic==1.13.0
annotated-types==0.7.0
anyio==3.7.1
arabic-reshaper==3.0.0
asn1crypto==1.5.1
asyncio==3.4.3
asyncpg==0.29.0
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
coverage==7.10.4
cryptography==45.0.6
cssselect2==0.8.0
distro==1.9.0
execnet==2.1.1
fastapi==0.104.1
frozenlist==1.7.0
greenlet==3.2.4
groq==0.4.1
gunicorn==21.2.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.2
logging==0.4.9.6
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.4
numpy==1.26.4
openai==1.3.7
opencv-python==4.8.1.78
oscrypto==1.3.0
packaging==25.0
Pillow==10.1.0
pluggy==1.6.0
prometheus-client==0.19.0
propcache==0.3.2
psutil==5.9.6
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
Pygments==2.19.2
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
pypdf==6.0.0
pytesseract==0.3.10
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-benchmark==4.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.3.1
python-bidi==0.6.6
python-decouple==3.8
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
PyYAML==6.0.2
redis==5.0.1
reportlab==4.0.9
requests==2.32.4
sentry-sdk==1.38.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
stripe==8.5.0
structlog==23.2.0
svglib==1.5.1
tinycss2==1.4.0
tqdm==4.67.1
twilio==8.10.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
xhtml2pdf==0.2.15
yarl==1.20.1
